Refaktorerer dokumenterte controllers i openapi til egne filer.
