+++ app/src/main/kotlin/no/nav/k9punsj/rest/web/openapi/OpenApi.kt
@@ -2,37 +2,17 @@ package no.nav.k9punsj.rest.web.openapi
 
 import com.fasterxml.jackson.annotation.JsonFormat
 import io.swagger.v3.core.converter.ModelConverter
-import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType
-import io.swagger.v3.oas.annotations.media.Content
-import io.swagger.v3.oas.annotations.media.Schema
-import io.swagger.v3.oas.annotations.responses.ApiResponse
-import io.swagger.v3.oas.annotations.responses.ApiResponses
-import io.swagger.v3.oas.annotations.security.SecurityRequirement
-import io.swagger.v3.oas.annotations.security.SecurityScheme
-import io.swagger.v3.oas.annotations.tags.Tag
 import io.swagger.v3.oas.models.OpenAPI
 import io.swagger.v3.oas.models.info.Contact
 import io.swagger.v3.oas.models.info.Info
 import io.swagger.v3.oas.models.info.License
 import io.swagger.v3.oas.models.servers.Server
-import no.nav.k9punsj.brev.BrevVisningDto
-import no.nav.k9punsj.brev.DokumentbestillingDto
 import no.nav.k9punsj.db.datamodell.MappeId
-import no.nav.k9punsj.fordel.FordelPunsjEventDto
 import no.nav.k9punsj.fordel.PunsjInnsendingType
-import no.nav.k9punsj.gosys.GosysRoutes
-import no.nav.k9punsj.journalpost.JournalpostRoutes
-import no.nav.k9punsj.rest.eksternt.pdl.PdlRoutes
-import no.nav.k9punsj.rest.web.IdentOgJournalpost
 import no.nav.k9punsj.rest.web.dto.*
-import no.nav.k9punsj.rest.web.ruter.BrevRoutes
-import no.nav.k9punsj.rest.web.ruter.OmsorgspengerRoutes
-import no.nav.k9punsj.rest.web.ruter.PleiepengerSyktBarnRoutes
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.stereotype.Component
-import org.springframework.web.bind.annotation.*
 import java.net.URI
 import java.time.LocalDate
 import java.time.LocalTime
@@ -68,241 +48,6 @@ internal class OpenApi {
                         .url("https://github.com/navikt/k9-punsj/blob/master/LICENSE")
                 )
         )
-}
-
-@RestController
-@Tag(name = "Pleiepenger sykt barn søknad", description = "Håndtering av papirsøknader")
-internal class PleiepengerSyktBarnSoknadController {
-    @GetMapping(PleiepengerSyktBarnRoutes.Urls.HenteMappe, produces = ["application/json"])
-    @Operation(
-        summary = "Henter mappen til en person som inneholder søknader.",
-        description = "Sendes NorskIdente til person som headere.",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Henter mappen til en person som inneholder alle søknader",
-                content = [Content(
-                    schema = Schema(
-                        implementation = PleiepengerSøknadDto::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun HenteMappe(@RequestHeader("X-Nav-NorskIdent") norskIdent: String) {
-    }
-
-    @GetMapping(PleiepengerSyktBarnRoutes.Urls.HenteSøknad, produces = ["application/json"])
-    @Operation(
-        summary = "Hente eksisterende mappe"
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Søknaden",
-                content = [Content(
-                    schema = Schema(
-                        implementation = SvarPsbDto::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Mappen finnes ikke"
-            )
-        ]
-    )
-    fun HenteSøknad(
-        @PathVariable("soeknad_id") søknadId: String,
-    ) {
-    }
-
-    @PutMapping(
-        PleiepengerSyktBarnRoutes.Urls.OppdaterEksisterendeSøknad,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Oppdatere en søknad i en eksisterende mappe.",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Innhold på søknader er oppdatert og søknadene er klare for innsending.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = PleiepengerSøknadDto::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun OppdatereSøknad(
-        @RequestBody søknad: PleiepengerSøknadDto,
-    ) {
-    }
-
-
-    @PostMapping(
-        PleiepengerSyktBarnRoutes.Urls.SendEksisterendeSøknad,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Sende inn søknad til behandling i saksbehsandlingssystemet.",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "202",
-                description = "Søknaden er lukket for endring og sendt til behandling.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = no.nav.k9.søknad.Søknad::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Innsending feilet grunnet mangler i søknaden.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = SøknadFeil::class
-                    )
-                )]
-            ), ApiResponse(
-                responseCode = "409",
-                description = "En eller flere journalposter har blitt sendt inn fra før",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasFeil::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "500",
-                description = "Hvis det feiler uventet på server",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasFeil::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun SendSøknad(
-        @RequestBody søknad: OasSendSøknad,
-    ) {
-    }
-
-    @PostMapping(
-        PleiepengerSyktBarnRoutes.Urls.ValiderSøknad,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Valider søknad mot k9-format sin kontrakt",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "202",
-                description = "Søknaden er valider ok.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = no.nav.k9.søknad.Søknad::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Innsending feilet grunnet mangler i søknaden.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = SøknadFeil::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "500",
-                description = "Hvis det feiler uventet på server",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasFeil::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun ValiderSøknad(
-        @RequestBody søknad: OasSendSøknad,
-    ) {
-    }
-
-    @PostMapping(
-        PleiepengerSyktBarnRoutes.Urls.NySøknad,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Starte en helt ny søknad",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "201",
-                description = "Opprettet en mappe, bunke og en tom søknad. Jobb videre mot søknadIden for å oppdatere søknaden.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = PleiepengerSøknadDto::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun NySøknad(
-        @RequestBody søknad: OasOpprettNySøknad,
-    ) {
-    }
-
-    @PostMapping(
-        PleiepengerSyktBarnRoutes.Urls.HentInfoFraK9sak,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Henter perioder som ligger i k9-sak",
-        description = "Henter perioder som ligger i k9-sak",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Henter siste pleiepengersøknad fra k9-sak og gjør den tilgjengelig for visning",
-                content = [Content(
-                    schema = Schema(
-                        implementation = PerioderDto::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Fant ingen gjeldene søknad"
-            )
-        ]
-    )
-    fun HentInfoFraK9sak(@RequestBody matchFagsak: OasMatchfagsak) {
-    }
 }
 
 // Disse klassene er nødvendige for å eksponere søknadsformatet, så lenge applikasjonen benytter userialisert json internt
@@ -354,7 +99,6 @@ data class OasPleiepengerSyktBarSoknadMappeSvar(
     )
 }
 
-
 data class OasSoknadsfeil(
     val mappeId: MappeId?,
     val feil: List<FeilDto>?,
@@ -364,317 +108,6 @@ data class OasSoknadsfeil(
         val feilkode: String?,
         val feilmelding: String?,
     )
-}
-
-@RestController
-@SecurityScheme(
-    name = "BearerAuth",
-    type = SecuritySchemeType.HTTP,
-    scheme = "bearer",
-    bearerFormat = "JWT"
-)
-@Tag(name = "Journalposter", description = "Håndtering av journalposter")
-internal class JournalpostController {
-    @PostMapping(
-        JournalpostRoutes.Urls.OmfordelJournalpost,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "204",
-                description = "Journalpost omfordelt"
-            ),
-            ApiResponse(
-                responseCode = "401",
-                description = "Ikke innlogget"
-            ),
-            ApiResponse(
-                responseCode = "403",
-                description = "Ikke tilgang til journalposten"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Journalpost eksisterer ikke"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Omfordele journalpost",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun OmfordelJournalpost(
-        @PathVariable("journalpost_id") journalpostId: String,
-        @RequestBody body: JournalpostRoutes.OmfordelingRequest,
-    ) {
-    }
-
-    @PostMapping(
-        JournalpostRoutes.Urls.HentJournalposter,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Hent journalposter som ikke er ferdig behandlet på person",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasJournalpostIder::class
-                    )
-                )]
-            ),
-        ]
-    )
-    @Operation(
-        summary = "Hent journalposter som ikke er ferdig behandlet på person",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun HentJournalposter(
-        @RequestBody body: OasIdentDto,
-    ) {
-    }
-
-    @GetMapping(JournalpostRoutes.Urls.JournalpostInfo, produces = ["application/json"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Liste med dokumenter som er i journalposten.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasJournalpostInfo::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Journalpost kan ikke håndteres i Punsj"
-            ),
-            ApiResponse(
-                responseCode = "401",
-                description = "Ikke innlogget"
-            ),
-            ApiResponse(
-                responseCode = "403",
-                description = "Ikke tilgang til journalposten"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Journalpost eksisterer ikke"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Hente informasjon om en journalpost",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun HenteJournalpostInfo(
-        @PathVariable("journalpost_id") journalpostId: String,
-    ) {
-    }
-
-    @PostMapping(JournalpostRoutes.Urls.SettPåVent, produces = ["application/json"])
-    @Operation(
-        summary = "Hente informasjon om en journalpost",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun SettPåVent(
-        @PathVariable("journalpost_id") journalpostId: String,
-        @RequestBody body: OasSøknadId,
-    ) {
-    }
-
-    @PostMapping(JournalpostRoutes.Urls.LukkJournalpost, produces = ["application/json"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Lukket ok",
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Journalpost eksisterer ikke"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Hente informasjon om en journalpost",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun LukkJournalpost(
-        @PathVariable("journalpost_id") journalpostId: String,
-    ) {
-    }
-
-    @PostMapping(JournalpostRoutes.Urls.SkalTilK9sak, produces = ["application/json"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "True når den skal til k9sak, false hvis den skal til infotrygd",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasSkalTilInfotrygdSvar::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "403",
-                description = "Ikke tilgang til journalposten"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Journalpost eksisterer ikke"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Sjekker om jornalposten må behandles av infotrygd",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun SkalTilK9Sak(
-        @RequestBody body: OasPunsjBolleDto,
-    ) {
-    }
-
-    @GetMapping(JournalpostRoutes.Urls.Dokument, produces = ["application/pdf"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Dokumentet."
-            ),
-            ApiResponse(
-                responseCode = "401",
-                description = "Ikke innlogget"
-            ),
-            ApiResponse(
-                responseCode = "403",
-                description = "Ikke tilgang til journalposten"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Journalpost eksisterer ikke"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Hente dokumentet",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun HenteDokument(
-        @PathVariable("journalpost_id") journalpostId: String,
-        @PathVariable("dokument_id") dokumentId: String,
-    ) {
-    }
-
-    @GetMapping(JournalpostRoutes.Urls.ResettInfoOmJournalpost, produces = ["application/json"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Når resett er gjennomført",
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Kan ikke endre på journalpost som har blitt sendt inn"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Journalpost eksisterer ikke"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Resette journalpost",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun ResettInfoOmJournalpost(
-        @PathVariable("journalpost_id") journalpostId: String,
-    ) {
-    }
-
-    @GetMapping(JournalpostRoutes.Urls.HentHvaSomHarBlittSendtInn, produces = ["application/json"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Gir hva som har blitt sendt inn for journalposten",
-                content = [Content(
-                    schema = Schema(
-                        implementation = no.nav.k9.søknad.Søknad::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Denne journalposten har ikke blitt sendt inn"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Journalpost eksisterer ikke"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Gir som svar tilbake hva som har blitt sendt inn på journalposten",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun HentHvaSomHarBlittSendtInn(
-        @PathVariable("journalpost_id") journalpostId: String,
-    ) {
-    }
-
-    @GetMapping(JournalpostRoutes.Urls.LukkJournalpostDebugg, produces = ["application/json"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Hvis journalposten har blitt lukket",
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Denne journalposten har allerede blitt lukket"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Journalpost eksisterer ikke i punsj"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Lukker en journalpost i k9-punsj og k9-los",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun LukkJournalpostDebugg(
-        @PathVariable("journalpost_id") journalpostId: String,
-    ) {
-    }
-
-    @PostMapping(JournalpostRoutes.Urls.JournalførPåGenerellSak, produces = ["application/json"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Hvis journalposten har blitt ferdigstilt på generell sak",
-            ),
-            ApiResponse(
-                responseCode = "500",
-                description = "Internal server error eller Doarkiv har fått Internal server error"
-            )
-        ]
-    )
-    @Operation(
-        summary = "Journalfører journalposten mot generell sak og ferdigstiller",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun JournalførPåGenerellSak(
-        @RequestBody body: IdentOgJournalpost,
-    ) {
-    }
 }
 
 data class OasDokumentInfo(
@@ -716,45 +149,6 @@ data class OasSkalTilInfotrygdSvar(
     val k9sak: Boolean,
 )
 
-@RestController
-@Tag(name = "Pdl", description = "Hent aktørid fra norsk ident")
-internal class PdlController {
-    @PostMapping(PdlRoutes.Urls.HentIdent, consumes = ["application/json"], produces = ["application/json"])
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Henter aktørid fra fnummer",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasPdlResponse::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "401",
-                description = "Ikke innlogget"
-            ),
-            ApiResponse(
-                responseCode = "403",
-                description = "Ikke tilgang til å slå opp personen"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Personen eksisterer ikke"
-            )
-        ]
-    )
-
-    @Operation(
-        summary = "Henter aktørid fra fnummer",
-        description = "Henter aktørid fra fnummer",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun Hentident(@RequestBody body: OasHentPerson) {
-    }
-}
-
 data class OasHentPerson(
     val norskIdent: NorskIdentDto,
 )
@@ -762,438 +156,3 @@ data class OasHentPerson(
 data class OasPdlResponse(
     val person: PdlPersonDto,
 )
-
-
-@RestController
-@Tag(name = "Gosys", description = "Opprett journalføringsoppgave")
-internal class GosysController {
-    @PostMapping(
-        GosysRoutes.Urls.OpprettJournalføringsoppgave,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Oppretter journalføringsoppgave for fnummer og journalpostid"
-            ),
-            ApiResponse(
-                responseCode = "401",
-                description = "Ikke innlogget"
-            ),
-            ApiResponse(
-                responseCode = "403",
-                description = "Ikke tilgang til å opprette journalføringsoppgave"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Eksisterer ikke"
-            ),
-
-            ApiResponse(
-                responseCode = "500",
-                description = "Eksisterer ikke",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasFeil::class
-                    )
-                )]
-            )
-        ]
-    )
-
-    @Operation(
-        summary = "Oppretter journalføringsoppgave",
-        description = "",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    fun OpprettJournalføringsoppgave(@RequestBody body: GosysRoutes.GosysOpprettJournalføringsOppgaveRequest) {
-    }
-
-    @GetMapping(GosysRoutes.Urls.Gjelder, produces = ["application/json"])
-    @Operation(
-        summary = "Hente gyldige gjelder-verdier. (Public endpoint)",
-        description = "Nøkkel-verdien brukes som 'gjelder' ved opprettelse av journalføringsoppgave."
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Gyldige gjelder-verdier med tekst"
-            )
-        ]
-    )
-    fun HenteGjelder() {
-    }
-}
-
-@RestController
-@Tag(name = "HendelseMottaker", description = "Prosesserer")
-internal class HendelseController {
-    @PostMapping(
-        "/prosesserHendelse/",
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Prosessert",
-                content = [Content(
-                    schema = Schema(
-                        implementation = FordelPunsjEventDto::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "401",
-                description = "Ikke innlogget"
-            ),
-            ApiResponse(
-                responseCode = "403",
-                description = "Ikke tilgang til å opprette journalføringsoppgave"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Eksisterer ikke"
-            )
-        ]
-    )
-
-    @Operation(summary = "Prosesser hendelse", description = "", security = [SecurityRequirement(name = "BearerAuth")])
-    fun ProsesserHendelse(@RequestBody body: FordelPunsjEventDto) {
-    }
-}
-
-@RestController
-@Tag(name = "Omsorgspenger søknad", description = "Håndtering av søknader av typen omsorgspenger")
-internal class OmsorgspengerSoknadController {
-    @GetMapping(OmsorgspengerRoutes.Urls.HenteMappe, produces = ["application/json"])
-    @Operation(
-        summary = "Henter data på person for omsorgspenger",
-        description = "Sendes NorskIdente til person som headere.",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Henter mappen til en person som inneholder alle søknader",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OmsorgspengerSøknadDto::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun HenteMappe(@RequestHeader("X-Nav-NorskIdent") norskIdent: String) {
-    }
-
-    @PostMapping(
-        OmsorgspengerRoutes.Urls.NySøknad,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Starte en helt ny søknad",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "201",
-                description = "Opprettet en mappe, bunke og en tom søknad. Jobb videre mot søknadIden for å oppdatere søknaden.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OmsorgspengerSøknadDto::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun NySøknad(
-        @RequestBody søknad: OasOpprettNySøknad,
-    ) {
-    }
-
-    @GetMapping(OmsorgspengerRoutes.Urls.HenteSøknad, produces = ["application/json"])
-    @Operation(
-        summary = "Hente eksisterende mappe"
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Søknaden",
-                content = [Content(
-                    schema = Schema(
-                        implementation = SvarOmsDto::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Mappen finnes ikke"
-            )
-        ]
-    )
-    fun HenteSøknad(
-        @PathVariable("soeknad_id") søknadId: String,
-    ) {
-    }
-
-    @PutMapping(
-        OmsorgspengerRoutes.Urls.OppdaterEksisterendeSøknad,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Oppdatere en søknad i en eksisterende mappe.",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Innhold på søknader er oppdatert og søknadene er klare for innsending.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OmsorgspengerSøknadDto::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun OppdatereSøknad(
-        @RequestBody søknad: OmsorgspengerSøknadDto,
-    ) {
-    }
-
-    @PostMapping(
-        OmsorgspengerRoutes.Urls.SendEksisterendeSøknad,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Sende inn søknad til behandling i saksbehsandlingssystemet.",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "202",
-                description = "Søknaden er lukket for endring og sendt til behandling.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = no.nav.k9.søknad.Søknad::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Innsending feilet grunnet mangler i søknaden.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = SøknadFeil::class
-                    )
-                )]
-            ), ApiResponse(
-                responseCode = "409",
-                description = "En eller flere journalposter har blitt sendt inn fra før",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasFeil::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "500",
-                description = "Hvis det feiler uventet på server",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasFeil::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun SendSøknad(
-        @RequestBody søknad: OasSendSøknad,
-    ) {
-    }
-
-    @PostMapping(
-        OmsorgspengerRoutes.Urls.ValiderSøknad,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Valider søknad mot k9-format sin kontrakt",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "202",
-                description = "Søknaden er valider ok.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = no.nav.k9.søknad.Søknad::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Innsending feilet grunnet mangler i søknaden.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = SøknadFeil::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "500",
-                description = "Hvis det feiler uventet på server",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasFeil::class
-                    )
-                )]
-            )
-        ]
-    )
-    fun ValiderSøknad(
-        @RequestBody søknad: OasSendSøknad,
-    ) {
-    }
-
-    @PostMapping(
-        OmsorgspengerRoutes.Urls.HentArbeidsforholdIderFraK9sak,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Henter arbeidsforholdIder som ligger i k9-sak",
-        description = "Henter arbeidsforholdIder som ligger i k9-sak",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Henter arbeidsforholdIder som ligger i k9-sak",
-                content = [Content(
-                    schema = Schema(
-                        implementation = ArbeidsgiverMedArbeidsforholdId::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Ingen treff"
-            )
-        ]
-    )
-    fun HentInfoFraK9sak(@RequestBody matchFagsak: OasMatchfagsakMedPeriode) {
-    }
-}
-
-@RestController
-@Tag(name = "Brev-bestilling", description = "Håndtering av brevbestillinger fra punsj til k9formidling via kafka")
-internal class BrevController {
-    @PostMapping(
-        BrevRoutes.Urls.BestillBrev,
-        consumes = ["application/json"],
-        produces = ["application/json"]
-    )
-    @Operation(
-        summary = "Sender en brevbestilling med gitt mal og innhold",
-        description = "Sender en brevbestilling med gitt mal og innhold",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Bestiller brev",
-                content = [Content(
-                    schema = Schema(
-                        implementation = DokumentbestillingDto::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Innsending feilet grunnet mangler i bestillingen.",
-                content = [Content(
-                    schema = Schema(
-                        implementation = OasFeil::class
-                    )
-                )]
-            ),
-        ]
-    )
-    fun BestillBrev(
-        @RequestBody bestilling: DokumentbestillingDto,
-    ) {
-    }
-
-    @GetMapping(BrevRoutes.Urls.HentAlleBrev, produces = ["application/json"])
-    @Operation(
-        summary = "Sender en brevbestilling med gitt mal og innhold",
-        description = "Sender en brevbestilling med gitt mal og innhold",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Bestiller brev",
-                content = [Content(
-                    schema = Schema(
-                        implementation = BrevVisningDto::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "Fant ingen brev",
-            )
-        ]
-    )
-    fun HentAlleBrev(
-        @PathVariable("journalpost_id") journalpostId: String,
-    ) {
-    }
-
-    @GetMapping(BrevRoutes.Urls.HentAktørId, produces = ["application/json"])
-    @Operation(
-        summary = "Sender en brevbestilling med gitt mal og innhold",
-        description = "Sender en brevbestilling med gitt mal og innhold",
-        security = [SecurityRequirement(name = "BearerAuth")]
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Gir aktørId",
-                content = [Content(
-                    schema = Schema(
-                        implementation = AktørIdDto::class
-                    )
-                )]
-            ),
-            ApiResponse(
-                responseCode = "400",
-                description = "Klarte ikke hent aktør for gitt fnr",
-            )
-        ]
-    )
-    fun HentAktørId(
-        @RequestHeader("X-Nav-NorskIdent") norskIdent: String
-    ) {
-    }
-}
